requirements.txt
requests~=2.31.0
Flask~=3.0.2
pillow~=10.2.0
python-dotenv~=1.0.1